Unit Tests for the Grid (and fixes)
Unit tests for the Grid to ensure that it's not broken. It only implements tests for an individual grid, and there is very likely to be errors involved if the game attempts to do things over multiple grids. However, the game should not need this so this should not become an issue.