Beginning of Game Initialisation
Started implementing Game Initialisation, with Registries to register PieceInfo and Rules for the game. The RuleBase Registry may be moved in the future given that Rules can be loaded once when the game starts, while the PieceInfo needs to reinitialise every game. (Once cards are implemented, PieceInfo will actively change as the game is played).

In addition, I have made it so the "Play Game" button opens up the board and places 5 different registered pieces on the board, with a default texture.