Validation Rules
These are rules for validating Actions. This allows for preventing pieces taking their own team, or moving on the same spot, while leaving the ability to remove these validation checks following the event.

I plan to add an ordering to these validation checks in the future, so that it's easier to control which order the rules run in (so that if there's a rule that cancels one another one out, it will be guaranteed to run afterwards).