Action dependencies
Made it so Actions can depend on one or more other Actions, where if one of the dependencies are invalid then that action will also be invalid. This allows for "slide" moves to invalidate further actions.

In the future, if there are any issues with the current implementation, I will alter the code so that actions take dependents into account (Actions that depend on the Action), and instead passes this "invalid" state downwards.