Improve Action Validation
Due to how it was programmed, it was difficult to have two rules interact when they are out of order. Therefore, the ActionBase tag system was altered so that rather than tags being added / removed one at a time for a single Action before the Action is removed, it stores all the tags and allows Verification of other Actions to add / remove these tags.

To do this, each Action now has their own tags.

If it's negative the tag isn't added, and when it's positive the tag is added.

This avoids the issue around actions being judged out of order when they remove or add tags to one another.