Action Request RPC
Added the code for clients to request the server to make Actions, which allows connected players to make actions.

Players can also only move pieces on their team. The implementation of this could be slightly improved, considering how the player numbers are read and stored.

Finally, the local play button works using an OfflineMultiplayerPeer instead of an ENetMultiplayerPeer, with both players set to the server so that the server controls both.