A couple fixes
Fixed a couple issues with the game. One being the image loading checking with FileAccess instead of ResourceLoader, meaning it wouldn't get the sprite when the game is built.

The other is that going to an offline lobby, and then to an online one doesn't reset the player number ids, so it now defaults them to -1 when the server is started.