Altered calculation of Actions
Altered Actions so that they are instead all calculated at the start of the round, and stored until the next turn stats. This is due to the fact that game needs to know if the King is in check or not, and the King's movement needs to be stopped if it tries to enter another piece's attack range. Doing Actions this way allows for the check to be made easily.

However, the method of implementation needs to be changed and improved as there is a pause each turn as it does this. However, from a quite check it is likely to be the LineMoveRule, given the number of actions it creates for each piece. (For the Queen alone, it is 112 Actions due to the ability to move 7 spaces in any direction, and being able to attack and move in each of these 7 spaces)

In addition, due to the change having AttackActions even if there's no pieces, the "AttackNeedsTargetRule" was added which makes AttackActions invalid if they have no target.