Generic Grid + Fixed Memory Leak
Make the Grid use a Generic so that in a follow up commit the Grid for the Pieces (and spaces) can be separated from the Actions. This should make it quicker to find pieces on certain cells. If I want to further improve the Grid system, it would be using a Dictionary with Vector2I keys for storing cells as it's direct access for the Cells, rather than need to loop through them each time.

However, in the process I found that there was a memory leak as the GodotObjects initialised for the C# objects were never freed. This has been fixed by changing them all to Nodes so that they could be added to the tree. This allows these nodes to be automatically freed once the GameController is freed from memory.