Fixed Check
Due to the checks for having the King in Check being in the NewTurn loop, any piece added after the King would not Check it and instead be able to take it. The checks for Check were moved to be outside this loop, in its own loop, to fix this.