Altered the Action code
Altered the action code so that rather than the Rules having access to an Array, they instead call a function in the Piece to add the Action to the list. Given that Rules do not have an ordering as of now, it's better to limit the Rules to only be able to add rules rather than removing them. It also simplifies the usage slightly.