Removed Grid Typing + Make Action Grid
Due to an issue with how casting works with generics and how I store the Grid and GridCell in the GridItem for easy access, I was unable to get the Grids to be typed as I wanted. Therefore, the typing for the Grid has been removed.

Given how the main Grid will be a mix of Pieces and Spaces at some point, it shouldn't be too much of an issue if there are no mistakes.