Fixed Check Verification
Check was broken, pieces could check the King with invalid AttackActions even though they are invalid / can't check. The no_check was using the "tags" variable in ActionBase, which should be used to tag the Action in the ActionRules. However, it should have been using a verification alternative, so that has been added and fixed Check.

In addition, it ignores the action if it's invalid.