Moved state of game to GameState class
Moved the state of the game to a class called "GameState". It is cloneable, so that the GameState can be recreated for simulating moves before they are made. This is done for the implementation of enforced check moves, and figuring out if the king is in checkmate if it's unable to move.