Implementation of Forced Check moves
If a move is attempted that puts the king in Check, or leaves it in Check, the move will fail and the server will inform the client that the action failed. It may be nice to have a message of the reason why it failed in the future, for more user feedback.

The way it does this is by cloning the GameState and acting out the attempted action. It then checks if the king is still in Check and, if it is, denies the action. Otherwise it accepts it.

The implementation could possibly be improved, given how the links are made between the actions. Actions are given ids when they are added to the piece, and Dictionaries are used to link id -> Action, and also id -> dependent actions. In addition, special Actions have can provide dictionaries with custom id links.