Improved Notices
If more than 5 notices were added, then the oldest notice would be deleted before its timer runs out. However, as it used a tree timer this resulted in an error because the Label object was no longer present, resulting in an error when it tried to remove it.

It has been fixed by giving the notices their own scene, so that the timer is freed when the notices are freed and not after.