AttackAction specificVictims fix
Changed specificVictims to be a single Piece reference. When checkmate / check is being checked for, removing a teammate from being attacked as a specific victim resulted in an error due to trying to edit the array while enumerating through it.

Rather than fixing it by looping down the indices, I found it a better idea to instead make it target a single victim, given the game no longer redoes all of the action calculations each turn, so the list wouldn't have the teammate in the next turn.