Threading improvements
Further improved Threading, fixing an issue with how the Tasks were written (allowing multiple threads to start, and result in access errors), preventing quits while the thread is active (as stopping it mid-work would require some refactoring of how the game processes turns) and added mutexes around the code to avoid clashing with the thread when accessing data.

In addition, added a check to avoid Clients having to calculate Check and Checkmate, given that the server already does so to allow / deny the player's actions or to signal a player has lost.

Currently uses GodotThread for the thread, due to issues I had trying to get it to work. However, it should be easy to change it back to a System.Threading Thread.