King Required
Added a check to make the King required on the board, otherwise the player will be guaranteed to be in checkmate. It also ensures that the player is unable to take moves which directly lose them their King piece.

This may be altered later, given that there are instances where a King should be able to be taken by the player on accident (and lead to a checkmate).