Fixed GodotThread warning
Due to the Threads not being closed properly by waiting for them, the thread is now only created once the first time a Task is started, and waits using a Semaphore when there are no Tasks available. Once the GameController is freed, it sets the loop to stop looping and waits for the thread to stop.

This avoids the warning so that the Thread is freed correctly.