Fixed issue with non-acting Actions
If the King had a Valid SlideAction, it would be considered a possible "move" that the King could take to get out of Check, given that it extends the MoveAction. However, it's not an action that can be taken, nor will it move the King out of the way of Check. Adding a check so that non-acting Actions are ignored for Check.