Fix to Offline "Action leads to Check." + Selecting non-visible Actions
In offline play, if the player tried to move a piece on the side that isn't currently playing, it would tell the player that "Action leads to Check.", even if it's not in check because it would return that the player is in check as the actions wouldn't act (as it's not their turn).  Fixed by making it return if the player is currently in check or not.

The player could also select non-displayed actions of pieces, even when they shouldn't be able to. This was fixed by making sure that the action is both valid and acting, otherwise SlideActions (which are valid) could be selected. Also, the game mutex is also used now due to it looking at the piece data.