Dedicated Server after game end
The dedicated server would not be able to reset after the game ended, due to being stuck in the game with the popup open (that displayed to all games). The game now unloads for all players when the game finishes, and unloads the scene for the dedicated server so that it's ready for the next one.