Check Fix for invalid AttackActions
Fixed an error in the Check code, as it tried to Check for all invalid attacks directly on the King, when only valid attacks should work. This resulted in Checkmate being possible when attacks on the King were invalid.

The same does not apply for spaces the King moves to, as the attack actions in these positions may be invalid.