Initial Card Implementation
Cards are implemented using Listeners that are setup to wait for certain GameEvents that occur. When they occur, the Cards are able to do things to the game, in addition to stopping the actions or waiting for the cards to do something before the event can complete.

Cards can be instantiated via adding a CardFactory for them to the registry. This factory creates a card, allowing the server to pass the card ids to the clients so that they can add the same card to their game more easily.

Cards are added outside of movement actions, so they don't have issues regarding networking (as the GDScript will deal with that side of it). However, for Cards that may need special networking, they will likely need to use Signals like the game uses elsewhere.