Verification Tags Fix
Fixed an issue with the Verify tags in Actions, as they weren't counting the number of additions and subtractions to know if a tag should be present or not. As a result, if the King was targeted by a LineMoveRule's AttackAction despite it being blocked by another Piece, it would be considered a valid Check due to "no_check" being removed on the AttackAction that targets the King (as that the MoveActions remove it from their Attack, and the Attacks further on the Line are checked later). By including the tag counts, "no_check" won't be removed as it won't go below 1 if something is hit before the King.