Server Lobby implementation
Added Server Lobbies which open new game instances as people create servers, and deletes them as they become unused. It's simply opening up dedicated servers, assigning a port to them and giving them to the players for it to work, as implementing it as multiple scenes in Godot would require rewriting a lot of the Lobby and GameManager code, given that they are both Autoloads and are used as Singletons.