Fixed another Check bug
LineMove attacks were able to move through the team's own king and "check" positions of the enemy King as if their King wasn't there. This is because these actions need to check beyond the King, if it's an enemy, but it didn't in this case because the team of the King wasn't taken into account. The movement of the King is now checked correctly.