Couple of Cards
Added a couple more cards to test the functionality.

ShuffleCard shuffles the position of all pieces of each team by their linkIds. If there is more than one teammate with a matching linkId, it's randomised between them. This does not yet work with the server.

SinglePieceArmyCard changes all of the pieces on the board but the King into a specific type of piece. This one is unlikely to be used further into the game, as it's not that enjoyable to play. Potentially changing it to "Change all X pieces into Y pieces" may work better, so that not all pieces on the board are affected.

The game now also has the Godot RandomNumberGenerator. Any randomness in the game will use this, as the Seed of the RandomNumberGenerator can be passed on to the Clones. This means that the randomness can be "predicted" to know how a move will end. If cards use randomness, but don't want the randomness to be taken into account, they can simply check if the GameState is a tempState or not.