Random Seed to Clients
The server now, when the game starts, gives the random seed to the Clients so that they all have the same randomness. This fixed the ShuffleCard not correctly working for the Client and Server in the same way. However, it would be good to add functionality so that the Client is given the server's Seed after each turn, given that there may be cards that only do Randomness on the Server. However, it may be better to program the cards in a way that avoids seed desync.