Altered PawnMoveRule
Changes "pawn_initial" to instead be removed when it's the piece's turn. This simplifies the tagging of it, as the reason for "setup_pawn_initial" was to change to "setup_initial" once the opponent's turn begins. By making it remove only on the piece's turn, it removes the need for this.