Moveed King MoveAction invalid in Check
Moved the check for Kings which make their MoveActions invalid if they are targeted by attacks to after the return point for if the GameState is a tempState, so that it doesn't have to be calculated in those instances. This is because the tempStates do not go further than checking 1 action further, and also that it's more for visualisation for the player, as the actions would be prevented anyway for putting the King into Check.