Tests for En Passant
Checks for En Passant to make sure that it's functioning correctly.

From this, a bug has been fixed where a Pawn that could be attacked via an En Passant was taken by another piece, then the Piece that was able to attack it would still be able to attack it as if it was there. This was fixed by updating a Pawn's possible actions in the following turn if it had started the turn next to another pawn, in addition to if it had ended the turn next to one.