Fixed King Check Test
The test originally checked if the King was in Check with the PlayerInCheck function, however I changed the CheckType of the player, when they have no king, to "NoKing". This has been fixed by changing it to use the new function "PlayerHasNoKing" where needed.

Also, changed the enum for the CheckType and CarryType to use CamelCase to be more consistent with the other enums.