CastleRule Tests
Implementation of tests for the CastleRule.

In the process, an issue was found with the Castling. The original implementation placed the King one before the Castle before jumping the Castle over, but the actual implementation of Castling should move the King 2 spaces from its starting location and then jump the Rook over one.