Moved game networking to GameManager
Moved the game networking that was in the game screen code to the GameManager, and connected them using signals. This allows for clients that join while the game is active to be in the lobby and not cause errors due to the game screen not existing for them to receive the rpcs.