Initial Card Implementation (without selection)
Implements the card selection from the deck, and addition to the game after a card is selected, which should work both offline and online. Actual card selection is not yet implemented, with the current implementation selecting the first card given by the server after 3 seconds of displaying them.

There is currently a memory access issue with C# when game screen is freed while the cards are being displayed, potentially due to the tweens not being freed correctly or the global timer awaits used in testing.