Single Piece Army Factory
To test a custom factory, the temporary "SinglePieceArmy" Card now randomly selects a piece that has been registered by the game to use. This works with clients loading the correct piece that the server selects via this factory.

The implementation uses gameRandom, the RandomNumberGenerator of the GameState, with the seed shared before the game starts, and therefore before the Cards activate. However, it may be better to make this and the Shuffle Card activate the moment they are added, as their effects change things the moment the game starts. This, however, will result in issues due to the random seed not being synced before the Major Card selection.