Card Selection
Card Selection was added, which gives cards for the users to select from once the game starts.

It was originally implemented in the GameManager, however due to the use of couroutines and the GameManager always being loaded the coroutine would continue waiting until a card has been selected, even if it's in a following game. This was fixed by moving it to its own Node.

The card visuals have been slightly improved too, so that the panels that make the card are not transparent.