Lonely Piece Stuck Card
Added a new Major Card that disallows all actions of a piece if it is not directly adjacent to another piece.

For this card, checking if a player has no actions available becomes more necessary, as it's much easier to reach a point where a player is unable to take any actions.