Cards activate on selection
Changed the SinglePieceArmyCard and ShuffleCard so that both activate when the card is added instead of when the game starts. This makes the changes more immediately visible, rather than only doing the changes when the game starts.

Due to this, the state of the random number generator needs to be sent from server to client after each card is selected, given that the server uses randomisation during card selection that clients don't.