Card hover visual fix
If a card was hovered over, selected and unhovered in quick succession, the card would visually appear halfway between it's up and down positions. This was due to the Tween being stopped, even if the card's move didn't occur due to it being held up. By moving this to the move functions, it has been fixed.