Updated Grid.
Finally changed Grid to use a Dictionary to store and retrieve Cells instead of an Array, improving the speed of the game processing drastically.

However, the Grid still uses the Array for the cell variable and finding Items on the Cells. This may be changed in the future, replacing the "cells" variable with an accessor to get an Array of the values, but for now it should be fine.