Modified King move invalidation + PieceDirection
Rather than checking for AttackActions, the game now checks the King's actual actions for what it can do. If an Action can get it out of Check, it can take the Action. This has also drastically sped up the DoesActionCheck function, and the GameState Clone function, as it no longer Clones *all* actions in the game anymore, and instead reuses the existing ones.

This does mean that future actions will require piece IDs to be used, rather than the Piece references, but for accurate Check / Checkmate finding that doesn't cause stuttering, it should be fine. As there are not many pieces on the board, the search for a specific id through the allPieces Array should not take too long. Though, making a dictionary to get pieces by their ids may be beneficial in the future.

In addition, Piece's forwardDirection is now an Enum so that it's easier to specify.