Move number of players to GameState
To avoid having any variables for the GameState in the GameController, the game now has its own numberOfPlayers variable that it uses for the number of players in the game. The only access the GameState does to the GameController now is for the Registries, but this will likely be changed in the future in favour of having the Registries in a separate class from the GameController, so that the GameState does not need to access the GameController at all.