Modifed Check move invalidation
Changed how the King's moves are invalidated in they put the piece in to Check by invalidating the Actions on that spot, rather than the actions and all of their dependents. This fixes an issue where a LineMoveRule piece can't move past the spaces that Check them, despite the fact that they should be able to.

This will likely need to be an option, as with this it's much harder to Checkmate pieces