Improved Piece Sprites
Improved the sprites of the Pieces, so that they can visually be told apart. The colouring was also changed to be white vs. black, rather than blue vs. red as it's the normal colouring for the pieces.

As the colouring uses shaders, it should be possible in the future to give user's control over what colours they want the pieces to be.