Separated Piece levels
Piece levels are now separated into individual levels, PieceInfo levels (for all pieces of one type) and Rule levels (for a single action of a piece type). These are summed when the the piece tries to get the actions of a specific rule.

Doing this allows for more control over the level of actions.