CastleRule fix
CastleRule actions would only disappear when the King moves. This means that as long as the King doesn't move, and the valid Castle actions exist, then the King will be able to Castle with the Rook even if it has moved.

This is now fixed by adding a tag to the King if it adds a castle action, and enabling actions update when the turn ends.