Made Actions sever-side
Due to how some cards will need to require user-input to function, it would be better to have all action processing and validation be server-side. The server then tells the clients to take certain actions as the server does them.

Because of this, when the cards need to pause, the clients will be at the same point as the server.