Lobby init fix
When game init started, there was a chance for a player to join mid-init, and have the rpcs for init_board or start_game to be called, despite not having init their games. This has been fixed by only having these rpcs run if the client was connected when init was started.