Initial implementation of PromotionCard
Adds the PromotionCard, which allows for one piece to promote to another piece when it reaches the end of the board. This card currently does not function if there are 2 or more possible pieces to promote to, as it will pause the game on the server-side while sending a notice to Godot to do the promotion.

In the future, a refactor of the game systems so that cards don't stop the code and instead use the Task system to repeat until it's done would work better. This would make the threading more resilient as a result.