Addons now require a card id
This is to prevent cards responding to signals they shouldn't and to allow more control over signals. There is another function provided which ignores this, and uses an empty array. With an empty array of card ids, it does not matter what card id sends the notice as it will always respond to the relevant notice.