Improved Piece selection
Piece selection now actively updates the action visuals when the actions are updated so that the player doesn't have to select the piece after the current actions taking place.

In addition, the hover highlight also hides while mid-action so that the user can more easily tell they can't take actions. It, however, doesn't disappear until the mouse moves to avoid the highlight flashing.