Fixed selection issue
When the player tried to play an invalid action (such as trying to move an enemy piece, or their piece when not their turn), it would result in the pieces resetting their action locations on the client. This results in the player being unable to request any actions, and the game getting stuck.

This is fixed by having it reset the piece actions after the first successful action on actions being sent to the client after actions start.