Slightly modified GameEvents wait
Rather than having the AnnounceEvents be separated into 2 functions, AnnounceEvents and AnnounceToListeners where AnnounceToListeners is an async Task, AnnounceEvents now calls the "Wait" async Task function when it needs to wait instead.

Due to some MinorCards only acting when added rather than when a GameEvent occurs, they need to be able to pause the game execution to wait for something. This allows for that.