Initial Minor Card Implementation
Minor Cards are added after the player builds up a "score" high enough to be given one. As of now, this is +1 per turn, and +2 for every piece lost.

Then, when the players have a high enough score AND it's their turn, they are given a selection of cards (currently only one). These are: Rule Cards, Piece Cards and Space Cards.

Rule Cards change a rule of the game. Currently, this is only the PromotionCard which changes a Piece into another Piece when it reaches the end of the board. However, rules for movement, attacks and other things will be added.

Piece Cards change a selected piece into the piece on the card. However, rather than changing the specific piece, it changes the linked piece too (This is normally the piece placed on the mirroring side).

Space Cards are not yet implemented, but will allow players to place special "spaces" which can do things when pieces land on them.