Lobby fix
When the game ends and the message pop up of the winner appears, this causes the game's processing to pause. Because of this, when the server owner leaves the player won't be kicked from the Lobby.

This resulted in a crash when pressing "Ok" as it would send the player to the Lobby screen and try to use the no longer existent multiplayer peer. Instead, the Lobby screen now kicks the user to the menu if there is no multiplayer peer when it loads, and so does the game screen.