Card Display
Upon selecting a card, it will show for all players at the top left of the screen for a second, before moving to be held somewhere. These places are the Major Card hand at the top of the screen, Player 1's hand at the bottom left and Player 2's at the bottom right.

The cards are also able to go off screen after being used, if they don't need to be displayed at all times.