ChangePiece fix
Moved the teams List add to before the validTeams List add, as otherwise it can block out teams from adding their id to the teams List. This would result in there being more teamIds in the validTeams List, even if it should be valid, not allowing for the card to be created.