Display card move fix
When selecting cards, if the user is viewing a card when the move_cards is called, it would put the card back into the display hand, despite the card still being in the "held" state. This resulted in the card getting stuck until another card is selected to be viewed.