Card description scroll
During card selection, the scroll bar will now slowly go up and down if the text is too big. This is not the case for display cards as it can be distracting, and the cards will display on the side when clicked.

It is not like this for card selection as card rise on hover there, making it harder to use the scroll bar, while the selected display card does not rise on hover.