Fix to a Piece taking itself
Instead of a Piece breaking the game by continuing to Act after taking itself, the game will now stop having the Piece act on finding it to be Taken or invalid (if freed early).

In addition, if a piece does not specify with an AttackAction to take itself, then it won't be taken.