Rock Piece
Adds a Rock Piece which can do nothing. Although this does not sound that useful, It works as a "skip turn" action. The Rock Piece can allow the player to use the turn on doing nothing, which means they can keep their pieces in an advantageous position if they're in one.

In addition, when Rule Cards are added, rules can be added to the Rocks which allow them to act in addition to skipping the turn.