ChangePiece on team pieces only
The biggest issue with the ChangePieceCard before was that you could select an enemy's piece you didn't like to swap them. With this change, you need to be more aware of which pieces are linked to the enemy pieces. In addition, if you lose a piece that the enemy has then you can no longer change that piece.