Improved LineRules
Modified the LineRules so that it's easier to create new ones, in addition to the directions being relative to the Piece, rather than set by the rule. The LineRuleBase still uses Vector2I, while the RelativeLineRuleBase uses this relative direction, in the case that direct positioning is required. However, it will likely go unused.