Piece viewing scene
A scene to allow for viewing the movement of pieces. This is more for screenshot taking rather than to be viewed in the game.

Use the "Run Current Scene" or "Run Specific Scene" buttons to run the scene.

The Left / Right arrows change the piece.

The Down / Up arrows change the timesMoved variable of the piece.

The - / = keys change the level of the piece.